Support Ubuntu 18.04
As Python 2 ends its lifetime in January 2020, the last Ubuntu LTS, which supports Python script in this repository is 18.04. Therefore I specified the preferred Ubuntu version to 18.04 and changed to use package(s) in this distribution. Also, to prevent a stall by the interactive input of the 'tzdata', set the install sequence to noninteractive mode.